Ensure HackRF is not sending out data during "STOP" mode
If the hackrf is not receiving any new data in-time, it will just repeat
sending the data which is still in its buffer. The copy block will
therefore not prevent the hackrf from sending anything when the "start"
pulse mode was selected before.

The flowgraph must therefore make sure that the buffer is filled with 0's
to make sure that the hackrf is not sending out any data.

Signed-off-by: Sven Eckelmann <sven@narfation.org>